rearrange and clean
