--- go-tilt-prototype/.gitignore
@@ -1,3 +1,0 @@
-bin/
-src/bin/
-src/obj/
