+++ go-tilt-prototype/basic-go-server/.gitignore
@@ -1,0 +1,1 @@
+build/
