+++ go-tilt-prototype/basic-go-server/go.mod
@@ -1,0 +1,5 @@
+module github.com/paketo-buildpacks/samples/go/mod
+
+go 1.15
+
+require github.com/gorilla/mux v1.8.0
