+++ go-tilt-prototype/basic-go-server/tilt_modules/extensions.json
@@ -1,0 +1,9 @@
+{
+  "Extensions": [
+    {
+      "Name": "pack",
+      "ExtensionRegistry": "https://github.com/tilt-dev/tilt-extensions",
+      "TimeFetched": "2021-10-07T10:21:28.200702-04:00"
+    }
+  ]
+}
