--- go-tilt-prototype/non-static/.gitignore
@@ -1,1 +1,0 @@
-build/
