--- go-tilt-prototype/non-static/go.mod
@@ -1,5 +1,0 @@
-module github.com/paketo-buildpacks/samples/go/mod
-
-go 1.15
-
-require github.com/gorilla/mux v1.8.0
