--- go-tilt-prototype/non-static/tilt_modules/extensions.json
@@ -1,9 +1,0 @@
-{
-  "Extensions": [
-    {
-      "Name": "pack",
-      "ExtensionRegistry": "https://github.com/tilt-dev/tilt-extensions",
-      "TimeFetched": "2021-10-07T10:21:28.200702-04:00"
-    }
-  ]
-}
