+++ go-tilt-prototype/server-with-template/.gitignore
@@ -1,0 +1,1 @@
+build/
