+++ go-tilt-prototype/server-with-template/go.mod
@@ -1,0 +1,8 @@
+module github.com/tilt-dev/tilt-example-go
+
+go 1.14
+
+require (
+	github.com/gorilla/mux v1.7.4
+	github.com/pkg/errors v0.9.1
+)
