+++ go-tilt-prototype/server-with-template/plan.toml
@@ -1,0 +1,6 @@
+[[requires]]
+  name = "watchexec"
+
+  [requires.metadata]
+    launch = true
+
