+++ go-tilt-prototype/server-with-template/web/templates/index.tmpl
@@ -1,0 +1,15 @@
+<!doctype html>
+<html>
+  <body style="font-size: 30px; font-family: sans-serif; margin: 0;">
+    <div style="display: flex; flex-direction: column; width: 100vw; height: 100vh; align-items: center; justify-content: center;">
+      <img src="pets.png" style="max-width: 30vw; max-height: 30vh;">
+      {{if .Congrats}}
+        <div>Congratulations, you hot reloaded a change!</div>
+      {{else}}
+        <div>Paketo + Go + Watchexec + Tilt, who knew?</div>
+      {{end}}
+
+      <div>{{.Greeting}} World</div>
+    </div>
+  </body>
+</html>
