--- go-tilt-prototype/static/deployments/Dockerfile
@@ -1,5 +1,0 @@
-FROM alpine
-WORKDIR /app
-ADD web web
-ADD build build
-ENTRYPOINT build/tilt-example-go
