--- go-tilt-prototype/static/go.mod
@@ -1,8 +1,0 @@
-module github.com/tilt-dev/tilt-example-go
-
-go 1.14
-
-require (
-	github.com/gorilla/mux v1.7.4
-	github.com/pkg/errors v0.9.1
-)
