--- go-tilt-prototype/static/start.go
@@ -1,5 +1,0 @@
-package main
-
-import "time"
-
-var StartTime = time.Unix(0, 1633548132559120640)
