--- go-tilt-prototype/static/web/templates/index.tmpl
@@ -1,15 +1,0 @@
-<!doctype html>
-<html>
-  <body style="font-size: 30px; font-family: sans-serif; margin: 0;">
-    <div style="display: flex; flex-direction: column; width: 100vw; height: 100vh; align-items: center; justify-content: center;">
-      <img src="pets.png" style="max-width: 30vw; max-height: 30vh;">
-      {{if .Congrats}}
-        <div>Congratulations, you set up live_update!</div>
-      {{else}}  
-        <div>Hello humans!</div>
-      {{end}}
-      
-      <div>Time from "deploy" button pressed → server up: {{.Duration}}</div>
-    </div>
-  </body>
-</html>
